[Add]: SchoolManagementApplication.java (class main) [Delete]: package com.example.studentmanagementsystem [Move]: move all folder about Student entity in package com.example.studentmanagementsystem to package no.restaff.schoolmanagement
