+++ src/main/java/no/restaff/schoolmanagement/SchoolManagementApplication.java
@@ -1,16 +1,16 @@
-package com.example.studentmanagementsystem;
+package no.restaff.schoolmanagement;
 
-import com.example.studentmanagementsystem.repository.StudentRepository;
+import no.restaff.schoolmanagement.repository.StudentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class StudentManagementSystemApplication implements CommandLineRunner {
+public class SchoolManagementApplication implements CommandLineRunner {
 
 	public static void main(String[] args) {
-		SpringApplication.run(StudentManagementSystemApplication.class, args);
+		SpringApplication.run(SchoolManagementApplication.class, args);
 	}
 
 	@Autowired
