+++ src/main/java/no/restaff/schoolmanagement/repository/StudentRepository.java
@@ -1,6 +1,6 @@
-package com.example.studentmanagementsystem.repository;
+package no.restaff.schoolmanagement.repository;
 
-import com.example.studentmanagementsystem.entity.Student;
+import no.restaff.schoolmanagement.entity.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StudentRepository extends JpaRepository<Student, Long> {
