+++ src/test/java/com/example/studentmanagementsystem/StudentManagementSystemApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class StudentManagementSystemApplicationTests {
+class SchoolManagementApplicationTests {
 
 	@Test
 	void contextLoads() {
